Refactor fat UI components into focused hooks and files (#1475)
Behavior-preserving refactor of four oversized components:

1. TaskPage: extract useTaskPageData hook — moves all useGrackle()
   destructuring, data-loading effects, derived state, and action
   bundles out of the page component (~200 lines → hook, leaving
   ~250 lines of JSX + URL/router state in the page).

2. TaskList: extract useTaskSearch + useGroupByStatus hooks and move
   StatusGroupAccordion + TaskTreeNode to their own files. Adds unit
   tests for both hooks.

3. EnvironmentEditPanel: move CodespacePicker to its own file and
   extract DockerContainerPicker from the inlined docker block.
   Adds Storybook stories for both extracted components.

4. WithSidebar: thin WithKnowledgeSidebar and WithCoordinationSidebar
   by extracting useKnowledgeSidebar and useCoordinationSidebar hooks.
   CoordinationOutletContext re-exported for backward compat.

Closes #1475
